refactor: split `lib/util.rs`
